third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_lineboundary_05.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'lineboundary'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} lineboundary`
  );
}

function makeSample(offset){
  const kSample = 'Lorem <span style="direction: rtl">ipsum dolor sit</span> amet';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right lineboundary
testIt('ltr', 0, 'right', '^Lorem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 1, 'right', 'L^orem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 2, 'right', 'Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 3, 'right', 'Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 4, 'right', 'Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 5, 'right', 'Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 6, 'right', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('ltr', 7, 'right', 'Lorem <^span style="direction: rtl">ipsum| dolor sit amet');
testIt('ltr', 8, 'right', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('ltr', 9, 'right', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('ltr', 10, 'right', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('ltr', 11, 'right', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('ltr', 12, 'right', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'right', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'right', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'right', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'right', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'right', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'right', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'right', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'right', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'right', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'right', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'right', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'right', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'right', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'right', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'right', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'right', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'right', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('ltr', 30, 'right', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('ltr', 31, 'right', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('ltr', 32, 'right', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('ltr', 33, 'right', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('ltr', 34, 'right', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 35, 'right', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('ltr', 36, 'right', 'Lorem <span style="direction: rtl">i^psum dolor sit</span> amet|');
testIt('ltr', 37, 'right', 'Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet|');
testIt('ltr', 38, 'right', 'Lorem <span style="direction: rtl">ips^um dolor sit</span> amet|');
testIt('ltr', 39, 'right', 'Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet|');
testIt('ltr', 40, 'right', 'Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet|');
testIt('ltr', 41, 'right', 'Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet|');
testIt('ltr', 42, 'right', 'Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet|');
testIt('ltr', 43, 'right', 'Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet|');
testIt('ltr', 44, 'right', 'Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet|');
testIt('ltr', 45, 'right', 'Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet|');
testIt('ltr', 46, 'right', 'Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet|');
testIt('ltr', 47, 'right', 'Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet|');
testIt('ltr', 48, 'right', 'Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet|');
testIt('ltr', 49, 'right', 'Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet|');
testIt('ltr', 50, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('ltr', 51, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet|</span>');
testIt('ltr', 52, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sitspan^ amet|</span>');
testIt('ltr', 53, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 54, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 55, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 56, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 57, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('ltr', 58, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet|');
testIt('ltr', 59, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met|');
testIt('ltr', 60, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et|');
testIt('ltr', 61, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t|');

// ltr left lineboundary
testIt('ltr', 0, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 1, 'left', '|L^orem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 2, 'left', '|Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 3, 'left', '|Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 4, 'left', '|Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 5, 'left', '|Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 6, 'left', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 7, 'left', '|Lorem <^span style="direction: rtl">ipsum dolor sit amet');
testIt('ltr', 8, 'left', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('ltr', 9, 'left', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('ltr', 10, 'left', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('ltr', 11, 'left', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('ltr', 12, 'left', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'left', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'left', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'left', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'left', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'left', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'left', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'left', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'left', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'left', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'left', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'left', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'left', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'left', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'left', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'left', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'left', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'left', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('ltr', 30, 'left', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('ltr', 31, 'left', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('ltr', 32, 'left', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('ltr', 33, 'left', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('ltr', 34, 'left', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 35, 'left', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 36, 'left', '|Lorem <span style="direction: rtl">i^psum dolor sit</span> amet');
testIt('ltr', 37, 'left', '|Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet');
testIt('ltr', 38, 'left', '|Lorem <span style="direction: rtl">ips^um dolor sit</span> amet');
testIt('ltr', 39, 'left', '|Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet');
testIt('ltr', 40, 'left', '|Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet');
testIt('ltr', 41, 'left', '|Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet');
testIt('ltr', 42, 'left', '|Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet');
testIt('ltr', 43, 'left', '|Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet');
testIt('ltr', 44, 'left', '|Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet');
testIt('ltr', 45, 'left', '|Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet');
testIt('ltr', 46, 'left', '|Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet');
testIt('ltr', 47, 'left', '|Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet');
testIt('ltr', 48, 'left', '|Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet');
testIt('ltr', 49, 'left', '|Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet');
testIt('ltr', 50, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('ltr', 51, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet</span>');
testIt('ltr', 52, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit^span amet</span>');
testIt('ltr', 53, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 54, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 55, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 56, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 57, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('ltr', 58, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet');
testIt('ltr', 59, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met');
testIt('ltr', 60, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et');
testIt('ltr', 61, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t');

// ltr forward lineboundary
testIt('ltr', 0, 'forward', '^Lorem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 1, 'forward', 'L^orem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 2, 'forward', 'Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 3, 'forward', 'Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 4, 'forward', 'Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 5, 'forward', 'Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('ltr', 6, 'forward', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('ltr', 7, 'forward', 'Lorem <^span style="direction: rtl">ipsum| dolor sit amet');
testIt('ltr', 8, 'forward', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('ltr', 9, 'forward', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('ltr', 10, 'forward', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('ltr', 11, 'forward', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('ltr', 12, 'forward', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'forward', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'forward', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'forward', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'forward', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'forward', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'forward', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'forward', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'forward', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'forward', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'forward', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'forward', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'forward', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'forward', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'forward', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'forward', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'forward', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'forward', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('ltr', 30, 'forward', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('ltr', 31, 'forward', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('ltr', 32, 'forward', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('ltr', 33, 'forward', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('ltr', 34, 'forward', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 35, 'forward', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('ltr', 36, 'forward', 'Lorem <span style="direction: rtl">i^psum dolor sit</span> amet|');
testIt('ltr', 37, 'forward', 'Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet|');
testIt('ltr', 38, 'forward', 'Lorem <span style="direction: rtl">ips^um dolor sit</span> amet|');
testIt('ltr', 39, 'forward', 'Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet|');
testIt('ltr', 40, 'forward', 'Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet|');
testIt('ltr', 41, 'forward', 'Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet|');
testIt('ltr', 42, 'forward', 'Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet|');
testIt('ltr', 43, 'forward', 'Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet|');
testIt('ltr', 44, 'forward', 'Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet|');
testIt('ltr', 45, 'forward', 'Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet|');
testIt('ltr', 46, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet|');
testIt('ltr', 47, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet|');
testIt('ltr', 48, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet|');
testIt('ltr', 49, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet|');
testIt('ltr', 50, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('ltr', 51, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet|</span>');
testIt('ltr', 52, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sitspan^ amet|</span>');
testIt('ltr', 53, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 54, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 55, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 56, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 57, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('ltr', 58, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet|');
testIt('ltr', 59, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met|');
testIt('ltr', 60, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et|');
testIt('ltr', 61, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t|');

// ltr backward lineboundary
testIt('ltr', 0, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 1, 'backward', '|L^orem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 2, 'backward', '|Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 3, 'backward', '|Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 4, 'backward', '|Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 5, 'backward', '|Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 6, 'backward', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 7, 'backward', '|Lorem <^span style="direction: rtl">ipsum dolor sit amet');
testIt('ltr', 8, 'backward', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('ltr', 9, 'backward', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('ltr', 10, 'backward', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('ltr', 11, 'backward', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('ltr', 12, 'backward', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 13, 'backward', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 14, 'backward', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 15, 'backward', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 16, 'backward', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 17, 'backward', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 18, 'backward', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('ltr', 19, 'backward', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 20, 'backward', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 21, 'backward', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 22, 'backward', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 23, 'backward', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 24, 'backward', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 25, 'backward', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 26, 'backward', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 27, 'backward', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 28, 'backward', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 29, 'backward', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('ltr', 30, 'backward', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('ltr', 31, 'backward', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('ltr', 32, 'backward', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('ltr', 33, 'backward', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('ltr', 34, 'backward', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 35, 'backward', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('ltr', 36, 'backward', '|Lorem <span style="direction: rtl">i^psum dolor sit</span> amet');
testIt('ltr', 37, 'backward', '|Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet');
testIt('ltr', 38, 'backward', '|Lorem <span style="direction: rtl">ips^um dolor sit</span> amet');
testIt('ltr', 39, 'backward', '|Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet');
testIt('ltr', 40, 'backward', '|Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet');
testIt('ltr', 41, 'backward', '|Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet');
testIt('ltr', 42, 'backward', '|Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet');
testIt('ltr', 43, 'backward', '|Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet');
testIt('ltr', 44, 'backward', '|Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet');
testIt('ltr', 45, 'backward', '|Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet');
testIt('ltr', 46, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet');
testIt('ltr', 47, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet');
testIt('ltr', 48, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet');
testIt('ltr', 49, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet');
testIt('ltr', 50, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('ltr', 51, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet</span>');
testIt('ltr', 52, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit^span amet</span>');
testIt('ltr', 53, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 54, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 55, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 56, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('ltr', 57, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('ltr', 58, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet');
testIt('ltr', 59, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met');
testIt('ltr', 60, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et');
testIt('ltr', 61, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t');

// rtl right lineboundary
testIt('rtl', 0, 'right', '^Lorem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 1, 'right', 'L^orem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 2, 'right', 'Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 3, 'right', 'Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 4, 'right', 'Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 5, 'right', 'Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 6, 'right', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('rtl', 7, 'right', 'Lorem <^span style="direction: rtl">ipsum| dolor sit amet');
testIt('rtl', 8, 'right', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('rtl', 9, 'right', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('rtl', 10, 'right', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('rtl', 11, 'right', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('rtl', 12, 'right', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'right', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'right', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'right', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'right', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'right', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'right', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'right', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'right', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'right', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'right', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'right', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'right', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'right', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'right', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'right', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'right', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'right', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('rtl', 30, 'right', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('rtl', 31, 'right', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('rtl', 32, 'right', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('rtl', 33, 'right', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('rtl', 34, 'right', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 35, 'right', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('rtl', 36, 'right', 'Lorem <span style="direction: rtl">i^psum dolor sit</span> amet|');
testIt('rtl', 37, 'right', 'Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet|');
testIt('rtl', 38, 'right', 'Lorem <span style="direction: rtl">ips^um dolor sit</span> amet|');
testIt('rtl', 39, 'right', 'Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet|');
testIt('rtl', 40, 'right', 'Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet|');
testIt('rtl', 41, 'right', 'Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet|');
testIt('rtl', 42, 'right', 'Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet|');
testIt('rtl', 43, 'right', 'Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet|');
testIt('rtl', 44, 'right', 'Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet|');
testIt('rtl', 45, 'right', 'Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet|');
testIt('rtl', 46, 'right', 'Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet|');
testIt('rtl', 47, 'right', 'Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet|');
testIt('rtl', 48, 'right', 'Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet|');
testIt('rtl', 49, 'right', 'Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet|');
testIt('rtl', 50, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('rtl', 51, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet|</span>');
testIt('rtl', 52, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sitspan^ amet|</span>');
testIt('rtl', 53, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 54, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 55, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 56, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 57, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('rtl', 58, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet|');
testIt('rtl', 59, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met|');
testIt('rtl', 60, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et|');
testIt('rtl', 61, 'right', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t|');

// rtl left lineboundary
testIt('rtl', 0, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 1, 'left', '|L^orem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 2, 'left', '|Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 3, 'left', '|Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 4, 'left', '|Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 5, 'left', '|Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 6, 'left', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 7, 'left', '|Lorem <^span style="direction: rtl">ipsum dolor sit amet');
testIt('rtl', 8, 'left', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('rtl', 9, 'left', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('rtl', 10, 'left', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('rtl', 11, 'left', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('rtl', 12, 'left', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'left', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'left', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'left', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'left', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'left', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'left', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'left', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'left', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'left', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'left', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'left', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'left', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'left', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'left', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'left', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'left', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'left', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('rtl', 30, 'left', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('rtl', 31, 'left', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('rtl', 32, 'left', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('rtl', 33, 'left', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('rtl', 34, 'left', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 35, 'left', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 36, 'left', '|Lorem <span style="direction: rtl">i^psum dolor sit</span> amet');
testIt('rtl', 37, 'left', '|Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet');
testIt('rtl', 38, 'left', '|Lorem <span style="direction: rtl">ips^um dolor sit</span> amet');
testIt('rtl', 39, 'left', '|Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet');
testIt('rtl', 40, 'left', '|Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet');
testIt('rtl', 41, 'left', '|Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet');
testIt('rtl', 42, 'left', '|Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet');
testIt('rtl', 43, 'left', '|Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet');
testIt('rtl', 44, 'left', '|Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet');
testIt('rtl', 45, 'left', '|Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet');
testIt('rtl', 46, 'left', '|Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet');
testIt('rtl', 47, 'left', '|Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet');
testIt('rtl', 48, 'left', '|Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet');
testIt('rtl', 49, 'left', '|Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet');
testIt('rtl', 50, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('rtl', 51, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet</span>');
testIt('rtl', 52, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit^span amet</span>');
testIt('rtl', 53, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 54, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 55, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 56, 'left', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 57, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('rtl', 58, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet');
testIt('rtl', 59, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met');
testIt('rtl', 60, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et');
testIt('rtl', 61, 'left', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t');

// rtl forward lineboundary
testIt('rtl', 0, 'forward', '^Lorem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 1, 'forward', 'L^orem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 2, 'forward', 'Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 3, 'forward', 'Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 4, 'forward', 'Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 5, 'forward', 'Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet|');
testIt('rtl', 6, 'forward', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('rtl', 7, 'forward', 'Lorem <^span style="direction: rtl">ipsum| dolor sit amet');
testIt('rtl', 8, 'forward', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('rtl', 9, 'forward', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('rtl', 10, 'forward', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('rtl', 11, 'forward', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('rtl', 12, 'forward', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'forward', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'forward', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'forward', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'forward', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'forward', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'forward', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'forward', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'forward', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'forward', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'forward', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'forward', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'forward', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'forward', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'forward', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'forward', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'forward', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'forward', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('rtl', 30, 'forward', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('rtl', 31, 'forward', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('rtl', 32, 'forward', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('rtl', 33, 'forward', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('rtl', 34, 'forward', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 35, 'forward', 'Lorem <span style="direction: rtl">^ipsum dolor sit</span> amet|');
testIt('rtl', 36, 'forward', 'Lorem <span style="direction: rtl">i^psum dolor sit</span> amet|');
testIt('rtl', 37, 'forward', 'Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet|');
testIt('rtl', 38, 'forward', 'Lorem <span style="direction: rtl">ips^um dolor sit</span> amet|');
testIt('rtl', 39, 'forward', 'Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet|');
testIt('rtl', 40, 'forward', 'Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet|');
testIt('rtl', 41, 'forward', 'Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet|');
testIt('rtl', 42, 'forward', 'Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet|');
testIt('rtl', 43, 'forward', 'Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet|');
testIt('rtl', 44, 'forward', 'Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet|');
testIt('rtl', 45, 'forward', 'Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet|');
testIt('rtl', 46, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet|');
testIt('rtl', 47, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet|');
testIt('rtl', 48, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet|');
testIt('rtl', 49, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet|');
testIt('rtl', 50, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('rtl', 51, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet|</span>');
testIt('rtl', 52, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sitspan^ amet|</span>');
testIt('rtl', 53, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 54, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 55, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 56, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 57, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span>^ amet|');
testIt('rtl', 58, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet|');
testIt('rtl', 59, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met|');
testIt('rtl', 60, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et|');
testIt('rtl', 61, 'forward', 'Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t|');

// rtl backward lineboundary
testIt('rtl', 0, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 1, 'backward', '|L^orem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 2, 'backward', '|Lo^rem <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 3, 'backward', '|Lor^em <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 4, 'backward', '|Lore^m <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 5, 'backward', '|Lorem^ <span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 6, 'backward', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 7, 'backward', '|Lorem <^span style="direction: rtl">ipsum dolor sit amet');
testIt('rtl', 8, 'backward', 'Lorem <s|pan style="direction: rtl">ipsum dolor sit amet</s|pan>');
testIt('rtl', 9, 'backward', 'Lorem <sp|an style="direction: rtl">ipsum dolor sit amet</sp|an>');
testIt('rtl', 10, 'backward', 'Lorem <spa|n style="direction: rtl">ipsum dolor sit amet</spa|n>');
testIt('rtl', 11, 'backward', 'Lorem <span| style="direction: rtl">ipsum dolor sit amet</span|>');
testIt('rtl', 12, 'backward', 'Lorem <span |style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 13, 'backward', 'Lorem <span s|tyle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 14, 'backward', 'Lorem <span st|yle="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 15, 'backward', 'Lorem <span sty|le="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 16, 'backward', 'Lorem <span styl|e="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 17, 'backward', 'Lorem <span style|="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 18, 'backward', 'Lorem <span rtl" style="|&quot;direction:">ipsum dolor sit</span> amet');
testIt('rtl', 19, 'backward', 'Lorem <span style="|direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 20, 'backward', 'Lorem <span style="d|irection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 21, 'backward', 'Lorem <span style="di|rection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 22, 'backward', 'Lorem <span style="dir|ection: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 23, 'backward', 'Lorem <span style="dire|ction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 24, 'backward', 'Lorem <span style="direc|tion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 25, 'backward', 'Lorem <span style="direct|ion: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 26, 'backward', 'Lorem <span style="directi|on: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 27, 'backward', 'Lorem <span style="directio|n: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 28, 'backward', 'Lorem <span style="direction|: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 29, 'backward', 'Lorem <span style="direction:| rtl">ipsum dolor sit</span> amet');
testIt('rtl', 30, 'backward', 'Lorem <span style="direction: |rtl">ipsum dolor sit</span> amet');
testIt('rtl', 31, 'backward', 'Lorem <span style="direction: r|tl">ipsum dolor sit</span> amet');
testIt('rtl', 32, 'backward', 'Lorem <span style="direction: rt|l">ipsum dolor sit</span> amet');
testIt('rtl', 33, 'backward', 'Lorem <span style="direction: rtl|">ipsum dolor sit</span> amet');
testIt('rtl', 34, 'backward', 'Lorem <span | style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 35, 'backward', '|Lorem ^<span style="direction: rtl">ipsum dolor sit</span> amet');
testIt('rtl', 36, 'backward', '|Lorem <span style="direction: rtl">i^psum dolor sit</span> amet');
testIt('rtl', 37, 'backward', '|Lorem <span style="direction: rtl">ip^sum dolor sit</span> amet');
testIt('rtl', 38, 'backward', '|Lorem <span style="direction: rtl">ips^um dolor sit</span> amet');
testIt('rtl', 39, 'backward', '|Lorem <span style="direction: rtl">ipsu^m dolor sit</span> amet');
testIt('rtl', 40, 'backward', '|Lorem <span style="direction: rtl">ipsum^ dolor sit</span> amet');
testIt('rtl', 41, 'backward', '|Lorem <span style="direction: rtl">ipsum ^dolor sit</span> amet');
testIt('rtl', 42, 'backward', '|Lorem <span style="direction: rtl">ipsum d^olor sit</span> amet');
testIt('rtl', 43, 'backward', '|Lorem <span style="direction: rtl">ipsum do^lor sit</span> amet');
testIt('rtl', 44, 'backward', '|Lorem <span style="direction: rtl">ipsum dol^or sit</span> amet');
testIt('rtl', 45, 'backward', '|Lorem <span style="direction: rtl">ipsum dolo^r sit</span> amet');
testIt('rtl', 46, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor^ sit</span> amet');
testIt('rtl', 47, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor ^sit</span> amet');
testIt('rtl', 48, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor s^it</span> amet');
testIt('rtl', 49, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor si^t</span> amet');
testIt('rtl', 50, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('rtl', 51, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit<^/span> amet</span>');
testIt('rtl', 52, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit^span amet</span>');
testIt('rtl', 53, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 54, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 55, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 56, 'backward', 'Lorem <span style="direction: rtl">ipsum dolor sit amet</span>');
testIt('rtl', 57, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit^</span> amet');
testIt('rtl', 58, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ^amet');
testIt('rtl', 59, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> a^met');
testIt('rtl', 60, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> am^et');
testIt('rtl', 61, 'backward', '|Lorem <span style="direction: rtl">ipsum dolor sit</span> ame^t');
</script>
